+++ apps/gjafakort/application/src/api/domains/audit/service.ts
@@ -1,0 +1,9 @@
+import AuditLog from './model'
+
+export const createAuditLog = (
+  state: string,
+  title: string,
+  authorSSN: string,
+  applicationId: string,
+  data: object = {},
+) => AuditLog.create({ state, title, authorSSN, applicationId, data })
